feat: add plugin
