+++ example/next.config.js
@@ -1,0 +1,5 @@
+const withRemoteRefresh = require('next-remote-refresh')({
+  paths: './package.json',
+  port: 4001,
+})
+module.exports = withRemoteRefresh()
