+++ example/package.json
@@ -3,7 +3,7 @@
   "version": "0.0.0",
   "private": "true",
   "scripts": {
-    "dev": "node ../bin/next-remote-refresh ../package.json --port=3002 & next dev"
+    "dev": "next dev"
   },
   "dependencies": {
     "next": "^11.0.1"
