+++ example/pages/index.js
@@ -1,11 +1,7 @@
 import path from 'path'
 import fs from 'fs'
-import { useRemoteRefresh } from 'next-remote-refresh'
 
 export default function Index({ name, version, description }) {
-  useRemoteRefresh({
-    port: 3002,
-  })
   return (
     <div>
       <h2>
